NODE-2600 State in utils/script/evaluate: * Balance is replaced by TxNonNegativeAmount; * Now an overridden balance can't be negative.
