+++ node/src/main/scala/com/wavesplatform/api/http/utils/UtilsInvocationRequest.scala
@@ -9,6 +9,7 @@ import com.wavesplatform.lang.ValidationError
 import com.wavesplatform.lang.directives.values.V6
 import com.wavesplatform.lang.v1.evaluator.ContractEvaluator.Invocation
 import com.wavesplatform.lang.v1.traits.domain.Recipient.Address as RideAddress
+import com.wavesplatform.state.BlockchainOverrides
 import com.wavesplatform.state.diffs.FeeValidation.{FeeConstants, FeeUnit}
 import com.wavesplatform.transaction.TxValidationError.GenericError
 import com.wavesplatform.transaction.smart.AttachedPaymentExtractor
